Add model to capture the status of upstream updates against an environment
